creating test parser for cache

+++ LRU_cache.h
@@ -5,12 +5,12 @@
 #include "list.h"
 
 typedef struct {
-    size_t capacity;
+    int capacity;
     List key_hits;
     Hash_map hash_table;
 }LRU_cache;
 
-void init_lru_cache(LRU_cache* cache, size_t capacity);
+void init_lru_cache(LRU_cache* cache, int capacity);
 int get_element(LRU_cache * cache, int key);
 void put_element(LRU_cache* cache, int key, int value);
